Proposal: change CSV format
Currently the CSV format is limited by:

* adding many information about the test setup into a single field,
  instead of using separate columns for the separate features
* doesn't include some parameters used during the run, such as the
  PG_TDE variable

This is just a draft/proposal, not an actual ready to be merged PR, as
it only hardcodes the PG_TDE variable instead of providing something
more generic, I just want to validate the idea with it first.

The usecase for this script is that a separete simple CSV processing
script can go through the CSV files, and upload them to a database used
by grafana, without:

* relying on information not existing in the csv, such as shell
  variables
* doing extensive string processing to try to recover all the
  information which was put into a single field

Toughts?